+++ CHANGES.md
@@ -1,4 +1,7 @@
 ## Changes in 0.6.1. (under development)
+* Removed constant-valued parameters from opener schema
+* Renamed store parameters `opensearch_url` and `opensearch_description_url` to
+  `endpoint_url` and `endpoint_description_url`, respectively.
 * Chunkstore considers bounding box when accessing data. Less data is accessed and normalized. (#33)
 * Fixed time range detection for datasets with daily time frequency.
 * Fixed problem with the encoding of a dataset's coordinate variables that occurs